+++ examples/hello-world/main.ts
@@ -1,10 +1,19 @@
-import { createRdfjsClient } from "@worlds/client/adapters/rdfjs";
+import { Client } from "@worlds/client";
+import { ComunicaSparqlEngine } from "@worlds/client/adapters/comunica";
+import {
+  RdfjsQuadStore,
+  RdfjsSearchIndex,
+} from "@worlds/client/adapters/rdfjs";
 import { QueryEngine } from "@comunica/query-sparql-rdfjs-lite";
+import { Store } from "n3";
 
 if (import.meta.main) {
+  const store = new Store();
   const queryEngine = new QueryEngine();
-  const client = createRdfjsClient({
-    queryEngine,
+  const client = new Client({
+    quadStore: new RdfjsQuadStore(store),
+    searchIndex: new RdfjsSearchIndex(store),
+    sparqlEngine: new ComunicaSparqlEngine({ queryEngine, store }),
   });
 
   await client.import({