+++ src/client/adapters/rdfjs/mod.ts
@@ -1,3 +1,2 @@
 export * from "./rdfjs-search-index.ts";
 export * from "./rdfjs-quad-store.ts";
-export * from "./create-rdfjs-client.ts";